+++ package.json
@@ -19,12 +19,10 @@
     "bootstrap": "^4.5.3",
     "chroma-js": "^2.1.0",
     "d3": "^6.2.0",
-    "history": "^4.10.1",
     "react": "^16.14.0",
     "react-bootstrap": "^1.4.0",
     "react-dom": "^16.14.0",
     "react-redux": "^7.2.1",
-    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
     "redux-saga": "^1.1.3",
     "styled-components": "^5.2.0",
@@ -52,7 +50,6 @@
     ]
   },
   "devDependencies": {
-    "@types/react-router-dom": "^5.1.6",
     "cross-env": "^7.0.2"
   }
 }
