+++ public/index.html
@@ -17,7 +17,8 @@
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <title>React App</title>
   </head>
-<body style="font-family: Piazzolla;">
+<body style="font-family: Piazzolla; background: #ffe5c8">
+    <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
   </body>
 </html>
