+++ src/BarPlot.tsx
@@ -1,13 +1,10 @@
-import React, { FC, useEffect, useContext } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { getBooks } from "./redux/actions/books";
+import React, { FC } from "react";
 import { OverlayTrigger, Popover } from "react-bootstrap";
 import * as d3 from "d3";
 import styled from "styled-components/macro";
 import chroma from "chroma-js";
-import { ThemeContext } from "./ThemeProvider";
 type BarPlotProps = {
-  category: any;
+  books: any;
 };
 const Wrapper = styled.div`
   div.arrow {
@@ -43,29 +40,8 @@ const Wrapper = styled.div`
     animation: swirl-in-fwd 0.6s ease-out both;
   }
 `;
-interface RootState {
-  books: {
-    books: Array<{
-      title: string;
-      rank: number;
-      rank_last_week: number;
-      weeks_on_list: number;
-    }>;
-    loading: boolean;
-    error: { message: string };
-  };
-}
-const BarPlot: FC<BarPlotProps> = ({ category }) => {
-  const { theme } = useContext(ThemeContext);
-  const dispatch = useDispatch();
-  const loading = useSelector((state: RootState) => state.books.loading);
-  const books = useSelector((state: RootState) => state.books.books);
-  const error = useSelector((state: RootState) => state.books.error);
 
-  useEffect(() => {
-    dispatch(getBooks(category));
-  }, [category, dispatch]);
-
+const BarPlot: FC<BarPlotProps> = ({ books }) => {
   const width = 600;
   const height = 560;
   const innerRadiusRanking = 60;
@@ -98,191 +74,184 @@ const BarPlot: FC<BarPlotProps> = ({ category }) => {
   function splitSentence(sentence: string) {
     let first = "";
     let second = "";
-    sentence.split(" ").map((word: string) => (
-      first.length > 15 ? (second += ` ${word}`) : (first += ` ${word}`)
-    ));
+    sentence.split(" ").map((word: string) => {
+      first.length > 15 ? (second += ` ${word}`) : (first += ` ${word}`);
+    });
     return { first, second };
   }
   return (
     <Wrapper>
-      {loading && <div>Loading...</div>}
-      {books.length > 0 && !loading && (
-        <svg
-          width={width}
-          height={height}
-          style={{ overflow: "visible", marginTop: "-1rem" }}
-        >
-          {books.map((d: any, idx: number) => {
-            const { first, second } = splitSentence(d.title);
-            return (
-              <OverlayTrigger
-                key={`bookPlot${idx}`}
-                trigger={["hover", "focus"]}
-                placement={idx > 7 ? "left" : "right"}
-                overlay={
-                  <Popover
-                    id="popover-basic"
-                    style={{
-                      borderBottomColor: "white",
-                      borderWidth: "3px",
-                      borderColor: "white",
-                      backgroundColor: `${color(idx)}`,
-                      maxWidth: "20rem",
-                      color: `white`,
-                    }}
-                  >
-                    <span className="d-flex p-3">
-                      <img alt="" height={150} src={d.book_image} />{" "}
-                      <span className="pl-3">
-                        <div
-                          style={{
-                            borderRadius: "0.2rem",
-                            padding: "0.2rem",
-                            marginBottom: "0.5rem",
-                            maxWidth: "6rem",
-                            backgroundColor: `${chroma(color(idx)!)
-                              .darken(0.5)
-                              .saturate(1)}`,
-                          }}
-                        >
-                          #{d.rank} this week
-                        </div>
-                        <strong>{d.title}</strong>
-                        <h6 className="font-italic">by {d.author}</h6>
-                        <div style={{ color: "#ffffffa8", fontSize: 12 }}>
-                          rank last week: #{d.rank_last_week}
-                        </div>
-                        <div>weeks on list: {d.weeks_on_list}</div>
-                      </span>
+      <svg
+        width={width}
+        height={height}
+        style={{ overflow: "visible", marginTop: "-1rem" }}
+      >
+        {books.map((d: any, idx: number) => {
+          const { first, second } = splitSentence(d.title);
+          return (
+            <OverlayTrigger
+              key={`bookPlot${idx}`}
+              trigger={["hover", "focus"]}
+              placement={idx > 7 ? "left" : "right"}
+              overlay={
+                <Popover
+                  id="popover-basic"
+                  style={{
+                    borderBottomColor: "white",
+                    borderWidth: "3px",
+                    borderColor: "white",
+                    backgroundColor: `${color(idx)}`,
+                    maxWidth: "20rem",
+                    color: `white`,
+                  }}
+                >
+                  <span className="d-flex p-3">
+                    <img alt="" height={150} src={d.book_image} />{" "}
+                    <span className="pl-3">
+                      <div
+                        style={{
+                          borderRadius: "0.2rem",
+                          padding: "0.2rem",
+                          marginBottom: "0.5rem",
+                          maxWidth: "6rem",
+                          backgroundColor: `${chroma(color(idx)!)
+                            .darken(0.5)
+                            .saturate(1)}`,
+                        }}
+                      >
+                        #{d.rank} this week
+                      </div>
+                      <strong>{d.title}</strong>
+                      <h6 className="font-italic">by {d.author}</h6>
+                      <div style={{ color: "#ffffffa8", fontSize: 12 }}>
+                        rank last week: #{d.rank_last_week}
+                      </div>
+                      <div>weeks on list: {d.weeks_on_list}</div>
                     </span>
-                  </Popover>
-                }
-              >
-                <g>
-                  <g
-                    key={`segment${idx}`}
-                    transform={`translate(${height / 2} ${width / 2})`}
-                  >
-                    <line
-                      transform={`rotate(${12 + 24 * idx})`}
-                      x1={0}
-                      y1={0}
-                      y2={300}
-                      stroke={"#E3CDA3"}
-                      strokeWidth={1}
-                      strokeDasharray={4}
-                    />
-                    <path
-                      className={`arcRank-${idx}`}
-                      d={
-                        createArc({
-                          startAngle: xScale(d.rank)!,
-                          endAngle: xScale(d.rank)! + bandWidth / 2,
-                          innerRadius: innerRadiusRanking,
-                          outerRadius: yScaleRanking(d.rank)!,
-                        })!
-                      }
-                      fill={color(idx)}
-                    />
-                    <path
-                      className={`arcRank-${idx}`}
-                      d={
-                        createArc({
-                          startAngle: xScale(d.rank)! + bandWidth / 2,
-                          endAngle: xScale(d.rank)! + bandWidth,
-                          innerRadius: innerRadiusRanking,
-                          outerRadius: yScaleRanking(
-                            d.rank_last_week === 0 ? 11 : d.rank_last_week
-                          )!,
-                        })!
-                      }
-                      fill={"#ffffffa8"}
-                    />
-                    <path
-                      className={`arcWeeks-${idx}`}
-                      d={
-                        createArc({
-                          startAngle: xScale(d.rank)!,
-                          endAngle: xScale(d.rank)! + bandWidth,
-                          innerRadius: innerRadiusWeeks,
-                          outerRadius: yScaleWeeks(
-                            d.weeks_on_list > 10 ? 10 : d.weeks_on_list
-                          )!,
-                        })!
-                      }
-                      fill={color(idx)}
-                    />
-                  </g>
-                  <g
-                    transform={`translate(${height / 2} ${
-                      width / 2
-                    }) rotate(170)`}
-                  >
-                    <text
-                      textAnchor="middle"
-                      y={innerRadiusRanking - 5}
-                      fontSize={11}
-                      transform={`rotate(${d.rank * 24})`}
-                      fill={theme.color}
-                    >
-                      {d.rank}
-                    </text>
-                  </g>
-                  <g
-                    transform={`translate(${height / 2} ${width / 2}) rotate(${
-                      175 + d.rank * 24
-                    })`}
-                  >
-                    // rotate around the chart
-                    <text
-                      textAnchor="start"
-                      y={outerRadiusWeeks}
-                      fontSize={11}
-                      transform={`rotate(90, ${xScale(
-                        d.rank
-                      )} ${outerRadiusWeeks})`} // rotate(a,x,y) rotate around a given point
-                    fill={theme.color}
-                    >
-                      {first}
-                    </text>
-                  </g>
-                  <g
-                    transform={`translate(${height / 2} ${width / 2}) rotate(${
-                      180 + d.rank * 24
-                    })`}
-                  >
-                    // rotate around the chart
-                    <text
-                      textAnchor="start"
-                      y={outerRadiusWeeks}
-                      fontSize={11}
-                      transform={`rotate(86, ${xScale(
-                        d.rank
-                      )} ${outerRadiusWeeks})`} // rotate(a,x,y) rotate around a given point
-                      fill={theme.color}
-                    >
-                      {second}
-                    </text>
-                  </g>
-                  <g
-                    transform={`translate(${height / 2} ${width / 2}) rotate(${
-                      170 + d.rank * 24
-                    })`}
-                  >
-                    <image
-                      y={outerRadiusWeeks}
-                      href={d.book_image}
-                      height="30"
-                      width="30"
-                    />
-                  </g>
+                  </span>
+                </Popover>
+              }
+            >
+              <g>
+                <g
+                  key={`segment${idx}`}
+                  transform={`translate(${height / 2} ${width / 2})`}
+                >
+                  <line
+                    transform={`rotate(${12 + 24 * idx})`}
+                    x1={0}
+                    y1={0}
+                    y2={300}
+                    stroke={"#E3CDA3"}
+                    strokeWidth={1}
+                    strokeDasharray={4}
+                  />
+                  <path
+                    className={`arcRank-${idx}`}
+                    d={
+                      createArc({
+                        startAngle: xScale(d.rank)!,
+                        endAngle: xScale(d.rank)! + bandWidth / 2,
+                        innerRadius: innerRadiusRanking,
+                        outerRadius: yScaleRanking(d.rank)!,
+                      })!
+                    }
+                    fill={color(idx)}
+                  />
+                  <path
+                    className={`arcRank-${idx}`}
+                    d={
+                      createArc({
+                        startAngle: xScale(d.rank)! + bandWidth / 2,
+                        endAngle: xScale(d.rank)! + bandWidth,
+                        innerRadius: innerRadiusRanking,
+                        outerRadius: yScaleRanking(
+                          d.rank_last_week === 0 ? 11 : d.rank_last_week
+                        )!,
+                      })!
+                    }
+                    fill={"#ffffffa8"}
+                  />
+                  <path
+                    className={`arcWeeks-${idx}`}
+                    d={
+                      createArc({
+                        startAngle: xScale(d.rank)!,
+                        endAngle: xScale(d.rank)! + bandWidth,
+                        innerRadius: innerRadiusWeeks,
+                        outerRadius: yScaleWeeks(
+                          d.weeks_on_list > 10 ? 10 : d.weeks_on_list
+                        )!,
+                      })!
+                    }
+                    fill={color(idx)}
+                  />
                 </g>
-              </OverlayTrigger>
-            );
-          })}
-        </svg>
-      )}
-      {error && <div>Failed to load data for {category} books.</div>}
+                <g
+                  transform={`translate(${height / 2} ${
+                    width / 2
+                  }) rotate(170)`}
+                >
+                  <text
+                    textAnchor="middle"
+                    y={innerRadiusRanking - 5}
+                    fontSize={11}
+                    transform={`rotate(${d.rank * 24})`}
+                  >
+                    {d.rank}
+                  </text>
+                </g>
+                <g
+                  transform={`translate(${height / 2} ${width / 2}) rotate(${
+                    175 + d.rank * 24
+                  })`}
+                >
+                  // rotate around the chart
+                  <text
+                    textAnchor="start"
+                    y={outerRadiusWeeks}
+                    fontSize={11}
+                    transform={`rotate(90, ${xScale(
+                      d.rank
+                    )} ${outerRadiusWeeks})`} // rotate(a,x,y) rotate around a given point
+                  >
+                    {first}
+                  </text>
+                </g>
+                <g
+                  transform={`translate(${height / 2} ${width / 2}) rotate(${
+                    180 + d.rank * 24
+                  })`}
+                >
+                  // rotate around the chart
+                  <text
+                    textAnchor="start"
+                    y={outerRadiusWeeks}
+                    fontSize={11}
+                    transform={`rotate(86, ${xScale(
+                      d.rank
+                    )} ${outerRadiusWeeks})`} // rotate(a,x,y) rotate around a given point
+                  >
+                    {second}
+                  </text>
+                </g>
+                <g
+                  transform={`translate(${height / 2} ${width / 2}) rotate(${
+                    170 + d.rank * 24
+                  })`}
+                >
+                  <image
+                    y={outerRadiusWeeks}
+                    href={d.book_image}
+                    height="30"
+                    width="30"
+                  />
+                </g>
+              </g>
+            </OverlayTrigger>
+          );
+        })}
+      </svg>
     </Wrapper>
   );
 };
