+++ src/PageWrapper.tsx
@@ -1,6 +1,5 @@
 import React, { FC } from "react";
 import styled from "styled-components/macro";
-
 const TypingEffect = `
   width:0;
   overflow: hidden; /* Ensures the content is not revealed until the animation */
@@ -21,9 +20,7 @@ const TypingEffect = `
   }
 `;
 const StyledPageWrapper = styled.div`
-background: ${({ theme }) => theme.backgroundColor};
-color: ${({ theme }) => theme.color};
-padding:2rem;
+  margin: 2rem;
   .table {
     float: right;
   }
@@ -82,7 +79,7 @@ padding:2rem;
  
 `;
 
-const PageWrapper: FC<{ theme: any }> = ({ theme, children }) => (
-  <StyledPageWrapper theme={theme}>{children}</StyledPageWrapper>
+const PageWrapper: FC = ({ children }) => (
+  <StyledPageWrapper>{children}</StyledPageWrapper>
 );
 export default PageWrapper;
