--- src/history.tsx
@@ -1,5 +1,0 @@
-import { createBrowserHistory } from 'history';
-
-const history = createBrowserHistory();
-
-export default history;
