+++ src/redux/reducers/types.js
@@ -1,11 +1,3 @@
 export const GET_BOOKS_REQUESTED = "GET_BOOKS_REQUESTED";
 export const GET_BOOKS_SUCCESS = "GET_BOOKS_SUCCESS";
 export const GET_BOOKS_FAILED = "GET_BOOKS_FAILED";
-
-export const LOGIN_USER = 'LOGIN_USER';
-export const LOGIN_USER_SUCCESS = 'LOGIN_USER_SUCCESS';
-export const LOGIN_USER_ERROR = 'LOGIN_USER_ERROR';
-
-export const GET_USERS_BOOKS_REQUESTED = 'GET_USERS_BOOKS_REQUESTED';
-export const GET_USERS_BOOKS_SUCCESS = 'GET_USERS_BOOKS_SUCCESS';
-export const GET_USERS_BOOKS_ERROR = 'GET_USERS_BOOKS_ERROR';
