+++ src/redux/sagas/index.js
@@ -1,8 +1,6 @@
 import { all } from 'redux-saga/effects'
 import booksSaga from "./booksSaga";
-import loginSaga from "./loginSaga";
-import usersBooksSaga from './usersBooksSaga';
 
 export default function* rootSaga() {
-  yield all([booksSaga(), loginSaga(), usersBooksSaga()]);
+  yield all([booksSaga()]);
 }
